1. remove docker network, too, during cleanup 2. moved test return value test from test_common_run_task_subprocess.py to test_common_run_task.py 3. small refactorings in docker_registry.py
