+++ setup.py
@@ -19,6 +19,7 @@ packages = \
  'exasol_integration_test_docker_environment.lib.docker.images.create.utils',
  'exasol_integration_test_docker_environment.lib.docker.images.push',
  'exasol_integration_test_docker_environment.lib.docker.images.save',
+ 'exasol_integration_test_docker_environment.lib.docker.networks',
  'exasol_integration_test_docker_environment.lib.docker.volumes',
  'exasol_integration_test_docker_environment.lib.logging',
  'exasol_integration_test_docker_environment.lib.test_environment',
